Update Makefile
Testing is a variable we can set now.
To run the program normally, use "make all"
To run the program as a test, use "make all TEST=1"
In testing mode, the header TEST_SYSTEM will be defined to all compiled files.

main.cpp takes advantage of this setup already by printing a message to the screen rather than opening up an Application instance


Former-commit-id: 9e7eb194b101769e7598e45c21a9246ab99ffc55