Console sucessfully creates games on server
Dialog class also created. To create a dialog use Dialog::show(GUIElement* parent, const std::string message); and forget about it. It stays on the screen until Enter or Escape is pressed.

TODO Finish the code for joining existing games. Pong seems to exit prematurely when it joins a game


Former-commit-id: e026d9a6138b482a003dd2b97bd957084bd94870