Configure Pong based on player ID
ID is set to the number of players in the game prior to the client joining that game, thus (if two people aren't simultaneously logging onto the same game) each player will have a unique ID in the game. Pong is set up to give the first player the left paddle, and it will give the second player the right paddle.

The only part left to implement is parsing the players' movement data coming from the server and controlling the paddles accordingly. This has not been done yet due to the server sending messages that are difficult to parse and will probably change in structure


Former-commit-id: 6163fe46a3fee2cd2a1601f9516967dd60554731